Tool for better initialization for transient
Add a tool function to compute a pluri-annual average value, which can be used as the first value of input series (recharge, flows for lakes...) in transient simulations. In this way, this average value will be used for the steady-state initialization of transient simulations, yielding to a more stable and relevant initial state.